Test merging - adding blank test files, renamed files and updated input_fragment.xml.
