+++ app/src/main/java/com/example/insertion_sort/InputFragment.java
@@ -1,0 +1,64 @@
+package com.example.insertion_sort;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.navigation.fragment.NavHostFragment;
+// Added 10/29/2024
+import android.text.Editable;
+import android.text.TextWatcher;
+import android.widget.Button;
+import android.widget.EditText;
+import androidx.annotation.Nullable;
+import androidx.navigation.Navigation;
+
+import com.example.insertion_sort.databinding.FragmentFirstBinding;
+
+public class InputFragment extends Fragment {
+
+    private FragmentFirstBinding binding;
+
+    @Nullable
+    @Override
+    public View onCreateView(
+            @NonNull LayoutInflater inflater, ViewGroup container,
+            Bundle savedInstanceState
+    ) {
+        // Inflate the layout for this fragment using View Binding
+        binding = FragmentFirstBinding.inflate(inflater, container, false);
+        return binding.getRoot();
+
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+
+        // Set up button click listener
+        binding.buttonSort.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                String inputText = binding.editTextArray.getText().toString().trim();
+                if (isInputValid(inputText)) {
+                    // Proceed with sorting logic here
+                    Toast.makeText(getActivity(), "Valid input: " + inputText, Toast.LENGTH_SHORT).show();
+                    // Call your sorting method here
+                } else {
+                    Toast.makeText(getActivity(), "Please enter a valid array of integers (0-9) with size between 3 and 8.", Toast.LENGTH_SHORT).show();
+                }
+            }
+        );
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        binding = null;
+    }
+
+}
