+++ app/src/main/res/layout/activity_main.xml
@@ -28,6 +28,13 @@
         android:layout_gravity="bottom|end"
         android:layout_marginEnd="@dimen/fab_margin"
         android:layout_marginBottom="16dp"
+        android:contentDescription="@string/fab_description"
         app:srcCompat="@android:drawable/ic_dialog_email" />
+
+    <FrameLayout
+        android:id="@+id/fragment_container"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:name="com.example.insertion_sort.InputFragment"/>
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
